Some changes in endpoint inputs and responses
