+++ FinanceServicesApi/Startup.cs
@@ -9,6 +9,7 @@ using FinanceServicesApi.V1.Boundary.Responses;
 using FinanceServicesApi.V1.Gateways;
 using FinanceServicesApi.V1.Gateways.Interfaces;
 using FinanceServicesApi.V1.Infrastructure;
+using FinanceServicesApi.V1.Infrastructure.Environments;
 using FinanceServicesApi.V1.Infrastructure.Interfaces;
 using FinanceServicesApi.V1.UseCase;
 using FinanceServicesApi.V1.UseCase.Interfaces;
