+++ FinanceServicesApi/V1/Domain/TenureModels/TenurePartialModel.cs
@@ -4,6 +4,12 @@ namespace FinanceServicesApi.V1.Domain.TenureModels
 {
     public class TenurePartialModel
     {
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <example>
+        ///     bdab4833-4079-420c-aeb0-b4f3611eae80
+        /// </example>
         public Guid Id { get; set; }
     }
 }
