+++ FinanceServicesApi/V1/Infrastructure/Environments/GetAccountEnvironmentVariables.cs
@@ -2,7 +2,7 @@ using System;
 using FinanceServicesApi.V1.Domain.AccountModels;
 using FinanceServicesApi.V1.Infrastructure.Interfaces;
 
-namespace FinanceServicesApi.V1.Infrastructure
+namespace FinanceServicesApi.V1.Infrastructure.Environments
 {
     public class GetAccountEnvironmentVariables : IGetEnvironmentVariables<Account>
     {
@@ -12,14 +12,6 @@ namespace FinanceServicesApi.V1.Infrastructure
             if (string.IsNullOrEmpty(result))
                 throw new Exception("Account api url shouldn't be null or empty.");
             return new Uri(result);
-        }
-
-        public string GetToken()
-        {
-            string result = Environment.GetEnvironmentVariable("ACCOUNT_API_TOKEN") ?? string.Empty; ;
-            if (string.IsNullOrEmpty(result))
-                throw new Exception("Account api token shouldn't be null or empty.");
-            return result;
         }
     }
 }
