+++ FinanceServicesApi/V1/Infrastructure/Interfaces/IGetEnvironmentVariables.cs
@@ -6,6 +6,5 @@ namespace FinanceServicesApi.V1.Infrastructure.Interfaces
     public interface IGetEnvironmentVariables<T> where T : class
     {
         public Uri GetUrl();
-        public string GetToken();
     }
 }
